requirements.txt
lightning==2.6.0
wandb==0.23.1
transformers==4.57.3
datasets==4.4.2
ruff==0.14.11
isort==7.0.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
